Cinvault/server/app/utils/object_id.py
from typing import Any

from bson import ObjectId
from fastapi import HTTPException, status


def to_object_id(value: str) -> ObjectId:
    if not ObjectId.is_valid(value):
        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Invalid id.")
    return ObjectId(value)


def serialize_doc(document: dict[str, Any] | None) -> dict[str, Any] | None:
    if document is None:
        return None

    serialized = dict(document)
    if "_id" in serialized:
        serialized["id"] = str(serialized.pop("_id"))

    for key, value in list(serialized.items()):
        if isinstance(value, ObjectId):
            serialized[key] = str(value)

    return serialized


def serialize_docs(documents: list[dict[str, Any]]) -> list[dict[str, Any]]:
    return [serialize_doc(document) for document in documents if document is not None]
