Cinvault/server/requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.35.0
motor==3.7.1
pymongo==4.13.2
pydantic-settings==2.10.1
python-dotenv==1.1.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
apscheduler==3.11.0
httpx==0.28.1
beautifulsoup4==4.12.3
email-validator==2.2.0
